handle not found

+++ inventory/src/main/java/br/edu/up/inventory/controllers/MovementController.java
@@ -2,6 +2,7 @@ package br.edu.up.inventory.controllers;
 
 import br.edu.up.inventory.domain.Movement;
 import br.edu.up.inventory.domain.MovementNature;
+import br.edu.up.inventory.domain.Process;
 import br.edu.up.inventory.domain.Product;
 import br.edu.up.inventory.domain.Warehouse;
 import br.edu.up.inventory.repository.MovementRepository;
@@ -17,6 +18,7 @@ import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
@@ -42,12 +44,20 @@ public class MovementController {
 
     @GetMapping("/{id}")
     Movement findById(@PathVariable Long id) {
-        return repository.findById(id).get();
+        Optional<Movement> optionalMovement =  repository.findById(id);
+        if (optionalMovement.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Movement not found");
+        }
+        return optionalMovement.get();
     }
 
     @PostMapping()
     Movement create(@RequestBody Movement newMovement) throws IOException, InterruptedException, ResponseStatusException {
-        Product product = productRepository.findById(newMovement.getIdProduct()).get();
+        Optional<Product> optionalProduct = productRepository.findById(newMovement.getIdProduct());
+        if (optionalProduct.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Product not found");
+        }
+        Product product = optionalProduct.get();
         int newQuantity = product.getQuantity();
         if (newMovement.getNature() == MovementNature.INCOMING) {
             newQuantity += newMovement.getQuantity();
@@ -114,6 +124,10 @@ public class MovementController {
 
     @DeleteMapping("/{id}")
     void delete(@PathVariable Long id) {
+        Optional<Movement> optionalMovement =  repository.findById(id);
+        if (optionalMovement.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Movement not found");
+        }
         repository.deleteById(id);
     }
